+++ package.json
@@ -28,6 +28,7 @@
     "@hapi/boom": "^9.1.0",
     "@hapi/cookie": "^11.0.1",
     "@hapi/hapi": "^19.1.1",
+    "@hapi/joi": "^17.1.1",
     "hapi-pino": "^8.0.0",
     "internal-ip": "^6.0.0",
     "migrate-mongo": "^7.2.1",
